fix: cfg

+++ particle/src/main/java/me/ibrahimsn/particle/Particle.kt
@@ -4,7 +4,7 @@ data class Particle (
     var radius: Float,
     var x: Float,
     var y: Float,
-    var vx: Int,
-    var vy: Int,
+    var vx: Float,
+    var vy: Float,
     var alpha: Int
 )
